requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
statsmodels>=0.14
scipy>=1.11
xgboost>=2.0
tensorflow==2.16.2
keras-tuner>=1.4
pyyaml>=6.0
tqdm>=4.65
matplotlib>=3.7
seaborn>=0.13
shap>=0.43
